+++ src/screens/userScreens/ManagementScreen.tsx
@@ -10,14 +10,13 @@ import {
   SafeAreaView,
   ScrollView,
 } from 'react-native';
-import {List} from 'react-native-paper';
 
 // Components 🪢
 import HeaderBar from '../../components/bannersAndBars/HeaderBar';
 import ToggleBar from './../../components/bannersAndBars/ToggleBar';
-import PollCard from '../../components/cards/PollCard';
 import CalendarManagement from '../../components/calendar/CalendarManagement';
 import {CoreButton} from '../../components/buttons/CoreButton';
+import PollsManagement from '../../components/screenComponents/PollsManagement';
 
 // StyleSheets 🌈
 import {CoreStyleSheet} from '../../StyleSheets/CoreDesignStyleSheet';
@@ -28,15 +27,11 @@ import color from '../../assets/defaultColorPallet.json';
 import {getLofftPolls} from '../../api/firebase/fireStoreActions';
 import firestore from '@react-native-firebase/firestore';
 
-// Helper Functions
-import dateCal from '../../components/helperFunctions/dateCal';
-
-const ManagementScreen = ({navigation, route}: any) => {
+const ManagementScreen = ({navigation}: any) => {
   // User Hooks
   const [image, setImage]: any = useState('');
   // Hooks
   const [pollsactivated, setPollsactivated] = useState(true);
-  const [expanded, setExpanded] = useState(true);
   const [date, setdate] = useState('');
   const [polls, setPolls] = useState([]);
   const [pastPolls, setPastPolls] = useState([]);
@@ -45,8 +40,6 @@ const ManagementScreen = ({navigation, route}: any) => {
   const buttonToggle = useCallback(toggled => {
     setPollsactivated(toggled);
   }, []);
-
-  const handlePress = () => setExpanded(!expanded);
 
   const fetchdate = dateInput => {
     setdate(dateInput);
@@ -92,10 +85,8 @@ const ManagementScreen = ({navigation, route}: any) => {
         CoreStyleSheet.viewContainerStyle,
         Platform.OS === 'ios' ? CoreStyleSheet.viewContainerIOSStyle : null,
       ]}>
-      <SafeAreaView style={styles.container}>
-        <ScrollView
-          style={styles.scrollView}
-          showsVerticalScrollIndicator={false}>
+      <SafeAreaView>
+        <ScrollView showsVerticalScrollIndicator={false}>
           <HeaderBar title="Management" image={image} />
 
           <ToggleBar
@@ -105,77 +96,11 @@ const ManagementScreen = ({navigation, route}: any) => {
           />
 
           {pollsactivated ? (
-            <>
-              <View style={styles.newPollContainer}>
-                <CoreButton
-                  value="Create New Poll"
-                  onPress={() => navigation.navigate('MakeNewPoll')}
-                  style={styles.newPollButton}
-                />
-              </View>
-
-              <List.Section>
-                <List.Accordion
-                  title={
-                    <View style={styles.accordionHeader}>
-                      <View style={styles.numberIcon}>
-                        <Text
-                          style={[
-                            fontStyles.bodySmall,
-                            {color: color.White[100]},
-                          ]}>
-                          {polls.length}
-                        </Text>
-                      </View>
-                      <Text style={fontStyles.buttonTextLarge}>
-                        Active Polls
-                      </Text>
-                    </View>
-                  }
-                  style={styles.accordionContainer}
-                  expanded={expanded}
-                  onPress={handlePress}>
-                  {polls.length > 0 ? (
-                    polls.map((el, index) => (
-                      <PollCard value={el} key={index} />
-                    ))
-                  ) : (
-                    <Text style={[fontStyles.bodyMedium, {marginLeft: 15}]}>
-                      Awkward, nothing here yet.... Create your first poll 🫀
-                    </Text>
-                  )}
-                </List.Accordion>
-              </List.Section>
-
-              <List.Section>
-                <List.Accordion
-                  title={
-                    <View style={styles.accordionHeader}>
-                      <View style={[styles.numberIcon, styles.pastPolls]}>
-                        <Text
-                          style={[
-                            fontStyles.bodySmall,
-                            {color: color.White[100]},
-                          ]}>
-                          {pastPolls.length}
-                        </Text>
-                      </View>
-                      <Text style={fontStyles.buttonTextLarge}>Past Polls</Text>
-                    </View>
-                  }
-                  style={styles.accordionContainer}>
-                  {pastPolls.length > 0 ? (
-                    pastPolls.map((el, index) => (
-                      <PollCard value={el} key={index} inactive />
-                    ))
-                  ) : (
-                    <Text style={[fontStyles.bodyMedium, {marginLeft: 15}]}>
-                      No Passt Polls yet
-                    </Text>
-                  )}
-                </List.Accordion>
-              </List.Section>
-            </>
+            <PollsManagement
+              navigation={navigation}
+              pastPolls={pastPolls}
+              polls={polls}
+            />
           ) : (
             <>
               <CalendarManagement fetchdate={fetchdate} />
@@ -195,45 +120,6 @@ const ManagementScreen = ({navigation, route}: any) => {
   );
 };
 
-const styles = StyleSheet.create({
-  newPollContainer: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    height: 60,
-    marginTop: 16,
-    paddingVertical: 8,
-  },
-  newPollButton: {
-    width: '60%',
-    flex: 1,
-  },
-  accordionContainer: {
-    backgroundColor: 'white',
-    display: 'flex',
-    justifyContent: 'flex-start',
-  },
-  accordionHeader: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  numberIcon: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: color.Mint[100],
-    width: 20,
-    height: 20,
-    borderRadius: 10,
-    marginRight: 5,
-  },
-  pastPolls: {
-    backgroundColor: color.Black[50],
-  },
-  buttonContainer: {
-    marginVertical: 120,
-  },
-  button: {
-    marginVertical: 5,
-  },
-});
+const styles = StyleSheet.create({});
 
 export default ManagementScreen;
